Redesigned core to support nanoseconds/data units, cgr-engine memory profiling in commandline options, tests updates

+++ engine/models.go
@@ -574,7 +574,7 @@ type SMCostSQL struct {
 	OriginHost  string
 	OriginID    string
 	CostSource  string
-	Usage       float64
+	Usage       int64
 	CostDetails string
 	CreatedAt   time.Time
 	DeletedAt   *time.Time
